@@ -5,26 +5,25 @@
   "private": true,
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "@types/xml": "^1.0.8",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.11",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
     "clsx": "^1.2.1",
     "lodash": "^4.17.21",
+    "npm-check-updates": "^16.6.2",
     "phaser": "^3.55.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-modal": "^3.16.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "sass": "^1.57.1",
-    "typescript": "^4.4.2",
+    "typescript": "^4.9.4",
     "use-interval": "^1.4.0",
-    "web-vitals": "^2.1.0"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -54,7 +53,6 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.191",
-    "@types/react-modal": "^3.13.1",
     "gh-pages": "^4.0.0",
     "stylelint-order": "^5.0.0",
     "stylelint-semantic-groups": "^1.2.0"
