@@ -1,4 +1,3 @@
-import PhaserGame from './phaser/PhaserGameConfig';
 import RadarScene from './phaser/scenes/RadarScene';
 
 import styles from './App.module.scss';
@@ -10,7 +9,6 @@ import {
 
 import React, { useEffect, useState } from 'react';
 import useInterval from 'use-interval';
-import Modal from 'react-modal';
 
 import './App.module.scss';
 import DepartureFDE from './react/components/DepartureFDE/DepartureFDE';
@@ -24,23 +22,13 @@ import clsx from 'clsx';
 import MenuSection from './react/components/MenuSection';
 import StripRow from './react/components/FdeSection';
 import FdeSection from './react/components/FdeSection';
-
-const examplePhaserFunction = () => {
-  const RADAR_SCENE = PhaserGame.scene.keys[
-    RadarSceneKeys.RADAR_06s
-  ] as RadarScene;
-  RADAR_SCENE.events.emit(ReactCustomEvents.CLICK_ME, { name: 'WOW' });
-};
+import { PhaserGameConfig } from './phaser/PhaserGameConfig';
 
 function App() {
-  const dispatch = useAppDispatch();
-
-  const airborneList = useAppSelector(selectDepartureList);
-
   return (
     <div className={styles.App}>
       <MenuSection appVersion="1.0" />
-      <FdeSection strips={airborneList} />
+      <FdeSection />
     </div>
   );
 }
