@@ -1,8 +1,7 @@
 import _ from 'lodash';
+import { AcType } from '../../phaser/types/AircraftTypes';
 import { RadarSceneKeys } from '../../phaser/types/SceneKeys';
 import { sidsCollection } from '../data/sidsCollection';
-
-import { AcType } from './genACID';
 
 export function genDepRoute(runwayId: RadarSceneKeys, acType: AcType) {
   if (!sidsCollection[runwayId]) return;
