@@ -1,8 +1,7 @@
 import _ from 'lodash';
+import { AcType } from '../../phaser/types/AircraftTypes';
 import { RadarSceneKeys } from '../../phaser/types/SceneKeys';
 import { satelliteCollection } from '../data/satelliteCollection';
-
-import { AcType } from './genACID';
 
 export function genSatRoute(runwayId: RadarSceneKeys, acType: AcType) {
   if (!satelliteCollection[runwayId]) return;
